requirements.txt
uvicorn
fastapi
Pillow
python-multipart
opencv-python